better handling of save file drop zone
handling is separated and dried
