+++ src/html/sidePanels.html
@@ -1,7 +1,5 @@
 <div id="loadSavePanel" class="sidePanel">
-    <div id="dropFileZone">
-        <p data-i18n="sidePanel:dropASaveFileHere">Drop a save file here</p>
-    </div>
+    <h2 data-i18n="sidePanel:loadSave">Load/Save</h2>
 </div>
 
 <div id="settingsPanel" class="sidePanel">
