+++ src/locales/en-US/sidePanel.json
@@ -1,4 +1,5 @@
 {
+    "loadSave": "Load a save",
     "dropASaveFileHere": "Drop a save file here",
     "settings": "Settings",
     "tutorials": "Tutorials",
