+++ src/locales/fr-FR/sidePanel.json
@@ -1,4 +1,5 @@
 {
+    "loadSave": "Charger une sauvegarde",
     "dropASaveFileHere": "Déposez un fichier de sauvegarde ici",
     "tutorials": "Tutoriels",
     "settings": "Paramètres",
