+++ src/styles/mainMenu.scss
@@ -98,14 +98,14 @@
   }
 
   #dropFileZone {
-    height: 100%;
-    width: 100%;
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
     cursor: pointer;
     transition: 0.1s;
+    border: 3px dashed grey;
+    margin: 0 10px;
 
     &.dragover {
       background: rgba(255, 255, 255, 0.1);
